+++ Source/BugsnagConfiguration.h
@@ -83,6 +83,15 @@ typedef void(^BugsnagOnSessionBlock)(NSMutableDictionary *_Nonnull sessionPayloa
 typedef NSDictionary *_Nullable (^BugsnagBeforeNotifyHook)(
     NSArray *_Nonnull rawEventReports, NSDictionary *_Nonnull report);
 
+typedef NS_OPTIONS(NSUInteger, BSGErrorType) {
+    BSGErrorTypesNone         = 0,
+    BSGErrorTypesOOMs         = 1 << 0,
+    BSGErrorTypesNSExceptions = 1 << 1,
+    BSGErrorTypesSignals      = 1 << 2,
+    BSGErrorTypesCPP          = 1 << 3,
+    BSGErrorTypesMach         = 1 << 4
+};
+
 @interface BugsnagConfiguration : NSObject
 
 /**
@@ -197,6 +206,13 @@ NSArray<BugsnagOnSessionBlock> *onSessionBlocks;
  * @see setEndpointsForNotify:sessions:
  */
 @property(readonly, retain, nullable) NSURL *sessionURL;
+
+/**
+ * A bitfield defining the types of error that are reported.
+ * Passed down to KSCrash in BugsnagCrashSentry.
+ * Defaults to all-true
+ */
+@property BSGErrorType enabledErrorTypes;
 
 /**
  * Required declaration to suppress a superclass designated-initializer error