+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/EnabledErrorTypesCxxScenario.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+#import "Scenario.h"
+
+@interface EnabledErrorTypesCxxScenario : Scenario
+@end